refactor(workbench): extract chat stage into WorkbenchChatStage
Move thread, stall/plan/find/tasks chrome, and error banners out of AppWorkbench so chat body can change without opening the host. Composer stays a child inside main__stage.
